requirements.txt
discord.py==2.3.2
requests==2.32.3
python-dotenv==1.0.1
flask==3.1.0
